Added new (experimental) docker build script for Apple Silicon chips.  Changed container launch script to have alternate commands for Mac and Git Bash (not tested yet, how it works on WSL bash). Readme updated

+++ run-container.sh
@@ -19,12 +19,15 @@ usage() {
 case "$(uname -s)" in
     Darwin*)    # macOS
         READLINK="greadlink"  # On macOS, use 'greadlink' from coreutils package
+        LAUNCHER="nix"
         ;;
     Linux*)     # Linux
         READLINK="readlink"
+        LAUNCHER="nix"
         ;;
     CYGWIN*|MINGW32*|MSYS*|MINGW*) # Windows
         READLINK="readlink"
+        LAUNCHER="gitbash"
         ;;
     *)
         >&2 echo "Error: Unsupported operating system."
@@ -63,6 +66,7 @@ if [[ $# -ne $((OPTIND)) ]]; then
     exit 1
 fi
 
+
 # Mount main project directory
 d=${@:$OPTIND:1}
 full_d=$($READLINK -f "${d}")
@@ -70,7 +74,13 @@ if [[ ! -d "${full_d}" ]]; then
   >&2 echo "Error: Directory ${full_d} does not exist."
   exit 1
 fi
-echo "${full_d} will be the main project mount directory"
+
+bname=$(basename ${full_d})
+if [[ "${bname}" == "geospaar"  ]]; then
+  >&2 echo "The mount directory cannot include geospaar, it should be"
+  >&2 echo "geospaar's parent directory"
+  exit 1
+fi
 
 # Make package directory
 r_package_dir=${full_d}/r_${ver}_packages
@@ -81,22 +91,36 @@ fi
 
 rstudio_image=agroimpacts/geospaar:${ver}
 check_image=$(docker image inspect ${rstudio_image})
-container_name=geospaar_rstudio
-
-# Running rstudio on localhost:8787
 prefs=rstudio-prefs.json
-docker run --rm -d -p 8787:${port} -e PASSWORD=password \
-  --name "${container_name}" \
-  -v "${full_d}":/home/rstudio/ \
-  -v "${r_package_dir}":/packages \
-  -v "${full_d}"/geospaar/"${prefs}":/home/rstudio/.config/rstudio/"${prefs}" \
-  -v "${full_d}"/geospaar/.Rprofile:/home/rstudio/.Rprofile:rw \
-  "${rstudio_image}"
 
-echo "${container_name} listening on port ${port}"
+# # Running rstudio on localhost:8787
+cd ${full_d}
+if [[ "${LAUNCHER}" == "nix"  ]]; then
+  echo Launching from a $LAUNCHER platform
+  docker run --rm -d -p 8787:${port} -e PASSWORD=password \
+    --name geospaar_rstudio \
+    -v "${full_d}":/home/rstudio/ \
+    -v "${r_package_dir}":/packages \
+    -v "${full_d}"/geospaar/"${prefs}":/home/rstudio/.config/rstudio/"${prefs}" \
+    -v "${full_d}"/geospaar/.Rprofile:/home/rstudio/.Rprofile:rw \
+    "${rstudio_image}"
+fi
+
+if [[ "${LAUNCHER}" == "gitbash"  ]]; then
+  echo Launching from a $LAUNCHER platform
+  winpty docker run --rm -d -p 8787:8787 -e PASSWORD=password \
+    --name geospaar_rstudio \
+    -v /$PWD:/home/rstudio/ \
+    -v /$PWD/r_$VER_packages:/packages \
+    -v /$PWD/geospaar/$prefs:/home/rstudio/.config/rstudio/$prefs \
+    -v /$PWD/geospaar/.Rprofile:/home/rstudio/.Rprofile:rw \
+    $rstudio_image
+fi
+
+echo "geospaar_rstudio listening on port ${port}"
 echo "Copy and paste http://localhost:${port} into your browser"
 echo "Username is rstudio and password is password"
-echo "To stop container run: docker stop ${container_name}"
+echo "To stop container run: docker stop geospaar_rstudio"
 # open http://localhost:${port}
